removes pepperjam.com as tracker
Pepperjam (Pepperjam.com) updated its public policy to make clear it does not engage in tracking as defined by Disconnect. Specifically, they state: “Pepperjam does not collect, retain or share any personally identifiable information gathered on sites or apps not owned by Pepperjam. This includes both unique user identifiers, as well as IP addresses, except for the limited purpose of fraud prevention and never to create user profiles in which case we collect and retain, but do not share, user IP for 12 months and never correlate a user IP across domains except when fraud is detected.”
https://www.pepperjam.com/legal#privacy